initial commit
This is an old project of mine that I created to enable network
communication between the Roomie desktop client and the Roomie web
server.  Nowadays it's generally best practice to use a REST API, but
this library has served me well so far.  The oldest file creation date
is October 2, 2010.  I pretty much banged the whole thing out in a week,
and then made some upgrades to create a strongly typed
TransmissionContext object in February 2012.